requirements/base.txt
Django==1.11.2
django-cors-headers==2.1.0
django-redis==4.8.0
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
PyJWT==1.5.2
pytz==2017.2
six==1.10.0
psycopg2==2.7.3.1
dj-database-url==0.4.2
